+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog for NeoFS Node
 - Iteration over locally collected container estimation values (#2597)
 - Synchronous objects `PUT` to the local storage (#2607)
 - `copies_number` field was not used in `PUT` request handling (#2607)
+- Neo-go notary deposit data workaround (#2429)
 
 ### Changed
 - FSTree storage now uses more efficient and safe temporary files under Linux (#2566)
@@ -30,7 +31,7 @@ Changelog for NeoFS Node
 
 ### Updated
 - Update minimal supported Go version up to v1.19 (#2485)
-- Update `neo-go` to `v0.102.0` (#2221, #2309, #2596)
+- Update `neo-go` to `v0.103.0` (#2221, #2309, #2596, #2626)
 - `neofs-lens` `inspect` object commands to `get` with `inspect` deprecation (#2548)
 - Update `tzhash` to `v1.7.1`
 - `github.com/panjf2000/ants/v2` to `v2.8.2`